+++ CHANGELOG.md
@@ -5,6 +5,7 @@ changes with their rationale when appropriate:
 
 ### v5.42.0.0
 - **http4k-*** : Upgrade some dependency versions.
+- **http4k-template** : Fix line separators for templates to allow working across platform with templates.
 - **http4k-connect-amazon-iamidentitycenter*** :[Fix 1239] CredentialsProvider.SSO - Credentials caching issue.
 - **http4k-connect-amazon-iamidentitycenter*** : [Unlikely break] PKCE-based Authorization for AWS SSO. Tweaks should be easy to remedy in previous SSO implementation. H/T @aharin